requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
pandas
sqllite
